refactor: 데이터 가공 역할 서비스 계층으로 이동
