+++ src/main/java/com/bbangle/bbangle/wishlist/repository/impl/WishListBoardRepositoryImpl.java
@@ -1,6 +1,5 @@
 package com.bbangle.bbangle.wishlist.repository.impl;
 
-import com.bbangle.bbangle.analytics.dto.AnalyticsCreatedWithinPeriodResponseDto;
 import com.bbangle.bbangle.analytics.dto.DateAndCountDto;
 import com.bbangle.bbangle.analytics.dto.QDateAndCountDto;
 import com.bbangle.bbangle.wishlist.repository.WishListBoardQueryDSLRepository;
@@ -12,12 +11,7 @@ import org.springframework.stereotype.Repository;
 
 import java.sql.Date;
 import java.time.LocalDate;
-import java.time.temporal.ChronoUnit;
 import java.util.List;
-import java.util.Map;
-import java.util.function.BiFunction;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import static com.bbangle.bbangle.wishlist.domain.QWishListBoard.wishListBoard;
 
@@ -29,12 +23,12 @@ public class WishListBoardRepositoryImpl implements WishListBoardQueryDSLReposit
 
 
     @Override
-    public AnalyticsCreatedWithinPeriodResponseDto countWishlistCreatedBetweenPeriod(LocalDate startLocalDate, LocalDate endLocalDate) {
+    public List<DateAndCountDto> countWishlistCreatedBetweenPeriod(LocalDate startLocalDate, LocalDate endLocalDate) {
         DateTemplate<Date> createdAt = getDateCreatedAt();
         Date startDate = Date.valueOf(startLocalDate);
         Date endDate = Date.valueOf(endLocalDate);
 
-        List<DateAndCountDto> rawResults = queryFactory.select(new QDateAndCountDto(
+        return queryFactory.select(new QDateAndCountDto(
                         createdAt, wishListBoard.id.count()
                 ))
                 .from(wishListBoard)
@@ -42,49 +36,11 @@ public class WishListBoardRepositoryImpl implements WishListBoardQueryDSLReposit
                 .groupBy(createdAt)
                 .orderBy(createdAt.asc())
                 .fetch();
-
-        List<DateAndCountDto> results = mapResultsToDateRangeWithCount(startLocalDate, endLocalDate, rawResults,
-                DateAndCountDto::date, DateAndCountDto::count, DateAndCountDto::new);
-
-        Long total = rawResults.stream()
-                .mapToLong(DateAndCountDto::count)
-                .sum();
-        Double daysBetween = calculateDaysBetween(startLocalDate, endLocalDate);
-
-        Double rawAverage = (total / daysBetween);
-        String average = String.format("%.2f", rawAverage);
-
-        return new AnalyticsCreatedWithinPeriodResponseDto(results, total, average);
     }
 
 
     private  DateTemplate<Date> getDateCreatedAt() {
         return Expressions.dateTemplate(Date.class, "DATE({0})", wishListBoard.createdAt);
-    }
-
-
-    public Double calculateDaysBetween(LocalDate startDate, LocalDate endDate) {
-        return (double) (ChronoUnit.DAYS.between(startDate, endDate) + 1);
-    }
-
-    private <T, R> List<R> mapResultsToDateRangeWithCount(
-            LocalDate startLocalDate, LocalDate endLocalDate, List<T> results,
-            Function<T, Date> dateExtractor,
-            Function<T, Long> countExtractor,
-            BiFunction<Date, Long, R> constructor) {
-
-        Map<Date, Long> mapResults = results.stream()
-                .collect(Collectors.toMap(
-                        dateExtractor,
-                        countExtractor
-                ));
-
-        List<LocalDate> dateRange = startLocalDate.datesUntil(endLocalDate.plusDays(1))
-                .toList();
-
-        return dateRange.stream()
-                .map(date -> constructor.apply(Date.valueOf(date), mapResults.getOrDefault(Date.valueOf(date), 0L)))
-                .toList();
     }
 
 }
